doc: Document all public types.
