+++ examples/keycode.rs
@@ -1,6 +1,5 @@
 use bevy::prelude::*;
-use bevy_input_sequence::*;
-use std::time::Duration;
+use bevy_input_sequence::prelude::*;
 
 #[derive(Clone, Debug)]
 enum Direction {
