+++ examples/keymod.rs
@@ -1,8 +1,5 @@
-use std::time::Duration;
-
 use bevy::prelude::*;
-
-use bevy_input_sequence::*;
+use bevy_input_sequence::prelude::*;
 
 #[derive(Event, Clone, Debug)]
 struct MyEvent;
