+++ src/action.rs
@@ -1,9 +1,7 @@
 //! Common actions to do on key sequence matches
 use bevy::ecs::{
     event::{Event, EventWriter},
-    schedule::Condition,
-    system::{In, IntoSystem, SystemId},
-    world::World,
+    system::{In},
 };
 
 /// Send this event.
